Split out masternodes list into it's own component
